backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
hypothesis==6.92.1
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
